chore(macOS): Merge element functionality with iOS
