+++ src/Uno.UI/UI/Xaml/FrameworkElement.iOS.cs
@@ -14,30 +14,6 @@ namespace Windows.UI.Xaml
 {
 	public partial class FrameworkElement
 	{
-		private bool _inLayoutSubviews;
-		private CGSize? _lastAvailableSize;
-		private CGSize _lastMeasure;
-
-		partial void Initialize();
-
-		internal bool RequiresArrange { get; private set; }
-
-		internal bool RequiresMeasure { get; private set; }
-
-
-		/// <summary>
-		/// Determines if InvalidateMeasure has been called
-		/// </summary>
-		/// <remarks>This property is present to mirror the WinUI souce</remarks>
-		internal bool IsMeasureDirty => RequiresMeasure;
-
-		/// <summary>
-		/// Determines if InvalidateArrange has been called
-		/// </summary>
-		/// <remarks>This property is present to mirror the WinUI souce</remarks>
-		internal bool IsArrangeDirty => RequiresArrange;
-
-
 		public override void SetNeedsLayout()
 		{
 			if (!_inLayoutSubviews)
@@ -49,11 +25,6 @@ namespace Windows.UI.Xaml
 			RequiresArrange = true;
 
 			SetSuperviewNeedsLayout();
-		}
-
-		public FrameworkElement()
-		{
-			Initialize();
 		}
 
 		public override void LayoutSubviews()
@@ -88,47 +59,6 @@ namespace Windows.UI.Xaml
 				this.Log().Error($"Layout failed in {GetType()}", e);
 			}
 		}
-		/// <summary>
-		/// Called before Arrange is called, this method will be deprecated
-		/// once OnMeasure/OnArrange will be implemented completely
-		/// </summary>
-		[EditorBrowsable(EditorBrowsableState.Never)]
-		protected virtual void OnBeforeArrange()
-		{
-
-		}
-
-		/// <summary>
-		/// Called after Arrange is called, this method will be deprecated
-		/// once OnMeasure/OnArrange will be implemented completely
-		/// </summary>
-		[EditorBrowsable(EditorBrowsableState.Never)]
-		protected virtual void OnAfterArrange()
-		{
-
-		}
-
-		private CGSize? XamlMeasure(CGSize availableSize)
-		{
-			// If set layout has not been called, we can 
-			// return a previously cached result for the same available size.
-			if (
-				!RequiresMeasure
-				&& _lastAvailableSize.HasValue
-				&& availableSize == _lastAvailableSize
-			)
-			{
-				return _lastMeasure;
-			}
-
-			_lastAvailableSize = availableSize;
-			RequiresMeasure = false;
-
-			var result = _layouter.Measure(SizeFromUISize(availableSize));
-
-			// Result here exclude the margins on the element
-			return _lastMeasure = result.LogicalToPhysicalPixels();
-		}
 
 		public override CGSize SizeThatFits(CGSize size)
 		{
@@ -151,28 +81,6 @@ namespace Windows.UI.Xaml
 			{
 				_inLayoutSubviews = false;
 			}
-		}
-
-		protected Size SizeFromUISize(CGSize size)
-		{
-			var width = nfloat.IsNaN(size.Width) ? float.PositiveInfinity : size.Width;
-			var height = nfloat.IsNaN(size.Height) ? float.PositiveInfinity : size.Height;
-
-			return new Size(width, height).PhysicalToLogicalPixels();
-		}
-
-		private bool IsTopLevelXamlView()
-		{
-			UIView parent = this;
-			while (parent != null)
-			{
-				parent = parent.Superview;
-				if (parent is IFrameworkElement)
-				{
-					return false;
-				}
-			}
-			return true;
 		}
 	}
 }
